requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
thefuzz==0.19.0
pandas==2.1.3
colorama==0.4.6
python-Levenshtein==0.23.0
pdfplumber==0.10.3
python-dateutil==2.8.2
werkzeug==3.0.1 